Initial wrapper around libfluent-bit
There are some problems with libfluent-bit that are not solved yet,
but this module can be used by first installing libfluent-bit

  curl https://raw.githubusercontent.com/fluent/fluent-bit/master/install.sh | sh

then setting

  export LD_PRELOAD=/lib/fluent-bit/libfluent-bit.so

then running "dzil test" or installing the module, etc.
The problem with the library is that it needs access to register
static threadlocal data before perl uses it all up.  (I found the
discussion of TLS static data and LD_PRELOAD workaround in various
discussions about OpenCV which aparently has similar problems)

The library paths are also hard-coded for now, but should be wrapped
in an Alien::LibFluentBit module to autodetect etc.  It might also
need to build a custom version of the library to avoid the static
TLS problem.

I'll save all that for later after we decide whether we're using this
for our logs.