Verious bug fixes, and started wrapping the library calls
The ability to pass pointers to flb_input, flb_filter and
flb_output was broken.  I have it correctly permitting undef
values for NULLs now.

then I wrapped the API in LibFluentBit with more perlish
methods, including tracking whether the engine is started
so that it can be cleanly shut down.

Then I wrote a simple logger object that writes timestamped
messages into the 'lib' input.

Added some unit tests, and it seems to be working.