Working unit test for datadog delivery
This code is still in a state of flux, but I finally have
the unit test successfully sending to datadog.  While the
API sequence is pretty much the same as documented, I was
running into a lot of weird behavior leading up to this,
so it might be dependent on which keys are found in the
JSON or how recent the timestamps are, or something.

The LD_PRELOAD is still required, for the moment.

I'm working on building a better perl-side API next.